+++ graph-resolver/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Graph resolver
 
+## 1.1.0
+
+13 March 2025
+
+* Fix issue where the same connection could be followed in both directions in one path
+* Fix issue with parallel paths not being followed (because a non-parallel part of
+  the path was already travelled)
+
 ## 1.0.0 - 1.0.1
 
 13 June 2024